favorites/steel - add api
