+++ basics/favorites/steel/api/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "favorites-steel-program"
+name = "favorites-steel-api"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,5 +7,7 @@ edition = "2021"
 crate-type = ["cdylib", "lib"]
 
 [dependencies]
+solana-program.workspace = true
 steel.workspace = true
-favorites-steel-api = {path = "../api"}
+bytemuck.workspace = true
+num_enum.workspace = true
