+++ basics/favorites/steel/api/src/consts.rs
@@ -1,0 +1,1 @@
+pub const FAVORITES_SEED: &[u8] = b"favorites";
